refactor(alarm_backends): 下沉Issue配置服务并统一Redis缓存读取
将 StrategyIssueConfigService 下沉到 models 层，并将 issue 配置读取链路调整为仅依赖 Redis。补充 issue 配置缓存的周期同步任务与写透/失效机制，确保后台处理不直接访问 DB。

Made-with: Cursor
