+++ bkmonitor/alarm_backends/core/cache/issue.py
@@ -8,51 +8,25 @@ an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express o
 specific language governing permissions and limitations under the License.
 """
 
-import json
 import logging
 
-from alarm_backends.core.storage.redis_cluster import RedisProxy
+from alarm_backends.core.cache.base import CacheManager
+from bkmonitor.utils import extended_json
 
 logger = logging.getLogger("fta_action.issue")
 
 STRATEGY_ISSUE_CONFIG_CACHE_TTL = 300
-STRATEGY_ISSUE_CONFIG_KEY_TPL = "issue.strategy.config.{strategy_id}"
 
 
-class StrategyIssueConfigCache:
+class StrategyIssueConfigCache(CacheManager):
     """StrategyIssueConfig 按 strategy_id 缓存，TTL 5min"""
 
-    _backend = "service"
+    CACHE_TIMEOUT = STRATEGY_ISSUE_CONFIG_CACHE_TTL
+    CACHE_KEY_TEMPLATE = CacheManager.CACHE_KEY_PREFIX + ".issue.strategy_config_{strategy_id}"
 
     @classmethod
-    def _get_client(cls):
-        return RedisProxy(cls._backend)
-
-    @classmethod
-    def _cache_key(cls, strategy_id: int) -> str:
-        return STRATEGY_ISSUE_CONFIG_KEY_TPL.format(strategy_id=strategy_id)
-
-    @classmethod
-    def get(cls, strategy_id: int):
-        """Redis → MySQL 降级 → 写缓存"""
-        client = cls._get_client()
-        cache_key = cls._cache_key(strategy_id)
-
-        cached = client.get(cache_key)
-        if cached:
-            try:
-                return json.loads(cached)
-            except (json.JSONDecodeError, TypeError):
-                pass
-
-        try:
-            from bkmonitor.models.issue import StrategyIssueConfig
-
-            config = StrategyIssueConfig.objects.get(strategy_id=strategy_id)
-        except Exception:
-            return None
-
-        data = {
+    def _serialize(cls, config) -> dict:
+        return {
             "strategy_id": config.strategy_id,
             "bk_biz_id": config.bk_biz_id,
             "is_enabled": config.is_enabled,
@@ -60,14 +34,64 @@ class StrategyIssueConfigCache:
             "conditions": config.conditions,
             "alert_levels": config.alert_levels,
         }
-        client.set(cache_key, json.dumps(data), ex=STRATEGY_ISSUE_CONFIG_CACHE_TTL)
+
+    @classmethod
+    def upsert(cls, config):
+        cache_key = cls.CACHE_KEY_TEMPLATE.format(strategy_id=config.strategy_id)
+        data = cls._serialize(config)
+        cls.cache.set(cache_key, extended_json.dumps(data), cls.CACHE_TIMEOUT)
         return data
+
+    @classmethod
+    def refresh(cls):
+        """全量同步 StrategyIssueConfig 到缓存，并清理已删除配置。"""
+        from bkmonitor.models.issue import StrategyIssueConfig
+
+        pipeline = cls.cache.pipeline()
+
+        configs = StrategyIssueConfig.origin_objects.filter(is_deleted=False).values(
+            "strategy_id",
+            "bk_biz_id",
+            "is_enabled",
+            "aggregate_dimensions",
+            "conditions",
+            "alert_levels",
+        )
+        for config in configs:
+            cache_key = cls.CACHE_KEY_TEMPLATE.format(strategy_id=config["strategy_id"])
+            pipeline.set(cache_key, extended_json.dumps(config), cls.CACHE_TIMEOUT)
+
+        deleted_strategy_ids = StrategyIssueConfig.origin_objects.filter(is_deleted=True).values_list(
+            "strategy_id", flat=True
+        )
+        for strategy_id in deleted_strategy_ids:
+            cache_key = cls.CACHE_KEY_TEMPLATE.format(strategy_id=strategy_id)
+            pipeline.delete(cache_key)
+
+        pipeline.execute()
+
+    @classmethod
+    def get(cls, strategy_id: int):
+        """仅从 Redis 获取配置。"""
+        cache_key = cls.CACHE_KEY_TEMPLATE.format(strategy_id=strategy_id)
+
+        cached = cls.cache.get(cache_key)
+        if cached:
+            try:
+                return extended_json.loads(cached)
+            except (TypeError, ValueError):
+                logger.warning("StrategyIssueConfigCache.get decode failed, strategy_id=%s", strategy_id)
+        return None
 
     @classmethod
     def invalidate(cls, strategy_id: int):
         """配置变更时主动删除缓存"""
         try:
-            client = cls._get_client()
-            client.delete(cls._cache_key(strategy_id))
+            cache_key = cls.CACHE_KEY_TEMPLATE.format(strategy_id=strategy_id)
+            cls.cache.delete(cache_key)
         except Exception:
             logger.warning("StrategyIssueConfigCache.invalidate failed, strategy_id=%s", strategy_id)
+
+
+def main():
+    StrategyIssueConfigCache.refresh()
